requirements.txt
python-dotenv>=1.0.0
langchain-core>=0.3.64
langchain-groq>=0.3.2
langchain-community>=0.3.24
langchain-text-splitters>=0.3.8
faiss-cpu>=1.7.3
sentence-transformers>=2.2.2
typing_extensions>=4.0.0
streamlit